// frontend/src/second/keeper-sqltree/types.ts
export interface Select {
  commandType: 'select';

  from: FromDefinition;

  columns?: ResultField[];
  topRecords?: number;
  range?: RangeDefinition;
  distinct?: boolean;
  selectAll?: boolean;
  orderBy?: OrderByExpression[];
  groupBy?: Expression[];
  where?: Condition;
  having?: Condition;
}
// ...
export interface UnaryCondition {
  expr: Expression;
}

export interface BinaryCondition {
  conditionType: 'binary';
  operator: '=' | '!=' | '<>' | '<' | '>' | '>=' | '<=';
  left: Expression;
  right: Expression;
}

export interface LikeCondition {
  conditionType: 'like' | 'notLike';
  left: Expression;
  right: Expression;
}

export interface NotCondition {
  conditionType: 'not';
  condition: Condition;
}

export interface TestCondition extends UnaryCondition {
  conditionType: 'isNull' | 'isNotNull' | 'isEmpty' | 'isNotEmpty';
}

export interface CompoudCondition {
  conditionType: 'and' | 'or';
  conditions: Condition[];
}

export interface ExistsCondition {
  conditionType: 'exists';
  subQuery: Select;
}
export interface NotExistsCondition {
  conditionType: 'notExists';
  subQuery: Select;
}

export interface BetweenCondition {
  conditionType: 'between';
  expr: Expression;
  left: Expression;
  right: Expression;
}

export interface InCondition {
  conditionType: 'in';
  expr: Expression;
  values: any[];
}

export interface RawTemplateCondition {
  conditionType: 'rawTemplate';
  templateSql: string;
  expr: Expression;
}

export type Condition =
  | BinaryCondition
  | NotCondition
  | TestCondition
  | CompoudCondition
  | LikeCondition
  | ExistsCondition
  | NotExistsCondition
  | BetweenCondition
  | InCondition
  | RawTemplateCondition;
// ...
export type Expression =
  | ColumnRefExpression
  | ValueExpression
  | PlaceholderExpression
  | RawExpression
  | CallExpression
  | MethodCallExpression
  | TranformExpression
  | RowNumberExpression;